@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
 import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { SectionService } from '../../../core/services/section.service';
+import { SectionService } from '../../../shared/services/section.service';
 
 @Injectable()
 export class SectionListResolver implements Resolve<boolean> {
