@@ -1,15 +1,15 @@
 import { Component } from '@angular/core';
 import { ActivatedRoute, PRIMARY_OUTLET, Router } from '@angular/router';
 import { DialogService } from 'primeng/dynamicdialog';
-import { NotificationsService } from '../../../core/services/notifications.service';
-import { MusicianProfileDto } from '../../../model/musicianProfileDto';
+import { NotificationsService } from '../../../../@arpa/services/notifications.service';
+import { MusicianProfileDto } from '../../../../@arpa/models/musicianProfileDto';
 import { TranslateService } from '@ngx-translate/core';
 import { MusicianLayoutComponent } from '../musician-layout/musician-layout.component';
 import { first, map, switchMap } from 'rxjs/operators';
-import { LoggerService } from '../../../core/services/logger.service';
-import { SectionDto } from '../../../model/sectionDto';
+import { SectionDto } from '../../../../@arpa/models/sectionDto';
 import { BehaviorSubject, combineLatest, of } from 'rxjs';
 import { MusicianService } from '../services/musician.service';
+import { LoggerService } from '../../../../@arpa/services/logger.service';
 
 @Component({
   selector: 'arpa-musician-dialog-entry',
