@@ -4,7 +4,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { TranslateMockModule } from '../../../../testing/translate.mock.module';
 import { ActivatedRoute } from '@angular/router';
 import { of } from 'rxjs';
-import { DummyComponent } from '../../../shared/dummy/dummy.component';
 import { DialogService } from 'primeng/dynamicdialog';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
@@ -20,7 +19,7 @@ describe('ProjectListComponent', () => {
           TranslateMockModule,
           HttpClientTestingModule,
         ],
-        declarations: [ProjectListComponent, DummyComponent],
+        declarations: [ProjectListComponent],
         providers: [
           { provide: DialogService, useValue: {} },
           { provide: ActivatedRoute, useValue: { data: of({}) } },
