@@ -3,8 +3,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { ProjectListComponent } from './project-list/project-list.component';
 import { ProjectGenreResolver } from './resolvers/project-genre.resolver';
 import { ProjectTypeResolver } from './resolvers/project-type.resolver';
-import { VenueService } from '../../core/services/venue.service';
-import { ProjectListResolver } from '../../core/resolvers/project-list.resolver';
+import { VenueService } from '../../shared/services/venue.service';
+import { ProjectListResolver } from '../../shared/resolvers/project-list.resolver';
 import { ProjectStateResolver } from './resolvers/project-status.resolver';
 
 const routes: Routes = [
